requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Web framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Rich output support
plotly>=5.14.0
dash>=2.14.0

# CLI
click>=8.1.0

# Optional dependencies
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Development
pytest>=7.3.0
pytest-asyncio>=0.21.0
black>=23.3.0
ruff>=0.0.270